requirements.txt
Django==2.2
django-bootstrap4==0.0.8
django-tables2==2.0.6
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
